+++ signin.js
@@ -1,5 +1,6 @@
-import * as signJs from "./sign.js";
-
+import { isValidEmail, alreadyLogin } from "./sign.js";
+import { ErrorMessage } from "./constant.js";
+import { signIn } from "./api.js";
 
 const myInputEmail = document.querySelector('.input-email');
 const myInputPassword = document.querySelector('.input-password');
@@ -19,44 +20,8 @@ const removeClassError = (errorMessage, currentClass) => {
   currentClass.classList.remove('error');
 };
 
+document.addEventListener('DOMContentLoaded', alreadyLogin);
 
-/*
-document.addEventListener('DOMContentLoaded', function() {
-  alreadyLogin();
-});
-*/
-
-
-function emailInputCheck() {
-  if (myInputEmail.value === '') {
-    addClassError(errorMessageId, myInputEmail, signJs.ErrorMessage.NoInputId);
-  } else if (!signJs.isValidEmail(myInputEmail.value)) {
-    addClassError(errorMessageId, myInputEmail, signJs.ErrorMessage.InvalidId);
-  } else {
-    removeClassError(errorMessageId, myInputEmail);
-  }
-}
-
-function passwordInputCheck() {
-  if (myInputPassword.value === '') {
-    addClassError(errorMessagePw, myInputPassword, signJs.ErrorMessage.NoInputPassword);
-  } else {
-    removeClassError(errorMessagePw, myInputPassword);
-  }
-}
-
-
-function emailInputValueCheck() {
-  signIn(myInputEmail.value, myInputPassword.value)
-  const answerEmail = 'test@codeit.com';
-  const answerPassword = 'codeit101';
-  if (myInputEmail.value === answerEmail && myInputPassword.value === answerPassword) {
-    window.location.href = '/folder';
-  } else {
-    addClassError(errorMessageId, myInputEmail, signJs.ErrorMessage.WrongId);
-    addClassError(errorMessagePw, myInputPassword, signJs.ErrorMessage.WrongPassword);
-  }
-}
 
 
 function toggleEyeImage() {
@@ -71,9 +36,27 @@ function toggleEyeImage() {
   }
 }
 
+function passwordInputCheckSignIn() {
+  if (myInputPassword.value === '') {
+    addClassError(errorMessagePw, myInputPassword, ErrorMessage.NoInputPassword);
+  } else {
+    removeClassError(errorMessagePw, myInputPassword);
+  }
+}
+
+function emailInputCheck() {
+  if (myInputEmail.value === '') {
+    addClassError(errorMessageId, myInputEmail, ErrorMessage.NoInputId);
+  } else if (!isValidEmail(myInputEmail.value)) {
+    addClassError(errorMessageId, myInputEmail, ErrorMessage.InvalidId);
+  } else {
+    removeClassError(errorMessageId, myInputEmail);
+  }
+}
+
+
 myInputEmail.addEventListener('focusout', emailInputCheck);
-myInputPassword.addEventListener('focusout', passwordInputCheck);
-/* login.addEventListener('click', emailInputValueCheck);*/
+myInputPassword.addEventListener('focusout', passwordInputCheckSignIn);
 
 // 검색하여 form 제출 시 페이지 새로고침되지 않게 함.
 /*
@@ -84,63 +67,6 @@ document.querySelector('.sign-form').addEventListener('submit', function(event) 
 */
 
 toggleButtonPw.addEventListener('click', toggleEyeImage);
-
-
-
-function signIn() {
-  const answerEmail = 'test@codeit.com';
-  const answerPassword = 'sprint101';
-
-  const data = {
-    email: answerEmail,
-    password: answerPassword
-  };
-
-  const inputData = {
-    email: myInputEmail.value,
-    password: myInputPassword.value
-  };
-
-  fetch('https://bootcamp-api.codeit.kr/api/sign-in', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify(data)
-  })
-  .then((response) => {
-    if (response.status === 200) {
-      return response.json();
-    }
-  })
-  .then((responseData) => {
-    console.log('서버 응답 데이터:', responseData);
-
-    // 서버 응답 데이터와 입력된 데이터 비교
-    if (data.email === inputData.email && data.password === inputData.password) {
-      localStorage.setItem("access-token", responseData.data.accessToken);
-      window.location.href = '/folder';
-    } else {
-      addClassError(errorMessageId, myInputEmail, signJs.ErrorMessage.WrongId);
-      addClassError(errorMessagePw, myInputPassword, signJs.ErrorMessage.WrongPassword);
-      console.log('여기서도 실행중');
-    }
-  })
-  .catch(error => {
-    console.error('요청 실패:', error);
-  });
-}
-
-/* function alreadyLogin() {
-  const accessToken = localStorage.getItem('access-token');
-  if (accessToken) {
-    window.location.href = '/folder';
-  }
-}
-*/
-
-
-
 
 export { 
   myInputEmail, 
@@ -154,6 +80,5 @@ export {
   addClassError,
   removeClassError,
   emailInputCheck,
-  passwordInputCheck,
   toggleEyeImage,
 };
